anged float to unsigned int in printf

+++ ncd.c
@@ -172,7 +172,7 @@ int comp_det()
 	setsockopt(send_fd, r, IP_TTL, &ttl, sizeof(ttl));
 
 	socklen_t size = 1500 * num_packets;
-	printf("Buffer size requested %f\n", (double)size);
+	printf("Buffer size requested %u\n", size);
 
 	setsockopt(send_fd, SOL_SOCKET, SO_SNDBUF, &size, sizeof(size));
